Clear term ui on Ctrl-C
